+++ src/DotNetty.Codecs.Http/WebSockets/Extensions/Compression/DeflateFrameClientExtensionHandshaker.cs
@@ -34,7 +34,7 @@ namespace DotNetty.Codecs.Http.WebSockets.Extensions.Compression
 
         public IWebSocketClientExtension HandshakeExtension(WebSocketExtensionData extensionData)
         {
-            if (extensionData.Parameters.Count > 0) { return null; }
+            if ((uint)extensionData.Parameters.Count > 0u) { return null; }
 
             var extensionDataName = extensionData.Name;
             switch (extensionDataName)
